+++ cpp/density/GaussianDensity.cc
@@ -46,73 +46,70 @@ void GaussianDensity::compute(const box::Box& box, const vec3<float>* points, un
     }
     m_density_array.prepare({width.x, width.y, width.z});
     util::ThreadStorage<float> local_bin_counts({width.x, width.y, width.z});
+
+    // set up some constants first
+    const float lx = m_box.getLx();
+    const float ly = m_box.getLy();
+    const float lz = m_box.getLz();
+
+    const float grid_size_x = lx / m_width.x;
+    const float grid_size_y = ly / m_width.y;
+    const float grid_size_z = m_box.is2D() ? 0 : lz / m_width.z;
+
+    // Find the number of bins within r_max
+    const int bin_cut_x = int(m_r_max / grid_size_x);
+    const int bin_cut_y = int(m_r_max / grid_size_y);
+    const int bin_cut_z = m_box.is2D() ? 0 : int(m_r_max / grid_size_z);
+
+    const float sigmasq = m_sigma * m_sigma;
+    const float A = std::sqrt(1.0f / (2.0f * M_PI * sigmasq));
+
     util::forLoopWrapper(0, n_points, [&](size_t begin, size_t end) {
-        // set up some constants first
-        float lx = m_box.getLx();
-        float ly = m_box.getLy();
-        float lz = m_box.getLz();
-
-        float grid_size_x = lx / m_width.x;
-        float grid_size_y = ly / m_width.y;
-        float grid_size_z = lz / m_width.z;
-
-        float sigmasq = m_sigma * m_sigma;
-        float A = sqrt(1.0f / (2.0f * M_PI * sigmasq));
-
         // for each reference point
         for (size_t idx = begin; idx < end; ++idx)
         {
-            // find the distance of that particle to bins
-            // will use this information to evaluate the Gaussian
-            // Find the which bin the particle is in
+            // Find which bin the particle is in
             int bin_x = int((points[idx].x + lx / 2.0f) / grid_size_x);
             int bin_y = int((points[idx].y + ly / 2.0f) / grid_size_y);
             int bin_z = int((points[idx].z + lz / 2.0f) / grid_size_z);
 
-            // Find the number of bins within r_max
-            int bin_cut_x = int(m_r_max / grid_size_x);
-            int bin_cut_y = int(m_r_max / grid_size_y);
-            int bin_cut_z = int(m_r_max / grid_size_z);
-
-            // in 2D, only loop over the 0 z plane
+            // In 2D, only loop over the z=0 plane
             if (m_box.is2D())
             {
                 bin_z = 0;
-                bin_cut_z = 0;
-                grid_size_z = 0;
             }
-            // Only evaluate over bins that are within the cut off
-            // to reduce the number of computations
+
+            // Only evaluate over bins that are within the cutoff
             for (int k = bin_z - bin_cut_z; k <= bin_z + bin_cut_z; k++)
             {
-                float dz = float((grid_size_z * k + grid_size_z / 2.0f) - points[idx].z - lz / 2.0f);
+                const float dz = float((grid_size_z * k + grid_size_z / 2.0f) - points[idx].z - lz / 2.0f);
 
                 for (int j = bin_y - bin_cut_y; j <= bin_y + bin_cut_y; j++)
                 {
-                    float dy = float((grid_size_y * j + grid_size_y / 2.0f) - points[idx].y - ly / 2.0f);
+                    const float dy = float((grid_size_y * j + grid_size_y / 2.0f) - points[idx].y - ly / 2.0f);
 
                     for (int i = bin_x - bin_cut_x; i <= bin_x + bin_cut_x; i++)
                     {
-                        // Calculate the distance from the grid cell to particular particle
-                        float dx = float((grid_size_x * i + grid_size_x / 2.0f) - points[idx].x - lx / 2.0f);
+                        // Calculate the distance from the particle to the grid cell
+                        const float dx = float((grid_size_x * i + grid_size_x / 2.0f) - points[idx].x - lx / 2.0f);
                         vec3<float> delta = m_box.wrap(vec3<float>(dx, dy, dz));
 
-                        float r_sq = dot(delta, delta);
-                        float r_sqrt = sqrtf(r_sq);
+                        const float r_sq = dot(delta, delta);
+                        const float r_sqrt = std::sqrt(r_sq);
 
                         // Check to see if this distance is within the specified r_max
                         if (r_sqrt < m_r_max)
                         {
                             // Evaluate the gaussian ...
-                            float x_gaussian = A * exp((-1.0f) * (delta.x * delta.x) / (2.0f * sigmasq));
-                            float y_gaussian = A * exp((-1.0f) * (delta.y * delta.y) / (2.0f * sigmasq));
-                            float z_gaussian = A * exp((-1.0f) * (delta.z * delta.z) / (2.0f * sigmasq));
+                            const float x_gaussian = A * exp((-1.0f) * (delta.x * delta.x) / (2.0f * sigmasq));
+                            const float y_gaussian = A * exp((-1.0f) * (delta.y * delta.y) / (2.0f * sigmasq));
+                            const float z_gaussian = A * exp((-1.0f) * (delta.z * delta.z) / (2.0f * sigmasq));
 
                             // Assure that out of range indices are corrected for storage
                             // in the array i.e. bin -1 is actually bin 29 for nbins = 30
-                            unsigned int ni = (i + m_width.x) % m_width.x;
-                            unsigned int nj = (j + m_width.y) % m_width.y;
-                            unsigned int nk = (k + m_width.z) % m_width.z;
+                            const unsigned int ni = (i + m_width.x) % m_width.x;
+                            const unsigned int nj = (j + m_width.y) % m_width.y;
+                            const unsigned int nk = (k + m_width.z) % m_width.z;
 
                             // store the product of these values in an array - n[i, j, k]
                             // = gx*gy*gz