+++ cpp/order/HexaticTranslational.cc
@@ -23,11 +23,10 @@ void HexaticTranslational<T>::computeGeneral(Func func, const freud::locality::N
 
         for(freud::locality::NeighborBond nb = ppiter->next(); !ppiter->end(); nb = ppiter->next())
         {
-            // Compute vector between the two particles
+            // Compute vector from query_point to point
             const vec3<float> delta = box.wrap((*points)[nb.point_idx] - ref);
 
-            // Compute psi for neighboring particle
-            // (only constructed for 2d)
+            // Compute psi for this vector
             m_psi_array[i] += func(delta);
         }